requirements.txt
nltk>=3.4.5
numpy>=1.16.4
gensim>=3.7.3
tensorflow>=2.2.0
recbole>=1.0.0
python>=3.7.7
pytorch>=1.7.1
faiss-gpu>=1.7.19
cudatoolkit>=10.1
